+++ src/main/kotlin/lotto/model/Rank.kt
@@ -1,4 +1,4 @@
-package lotto
+package lotto.model
 
 enum class Rank(val countOfMatch: Int, val winningMoney: Int) {
     FIRST(6, 2_000_000_000),
@@ -6,7 +6,8 @@ enum class Rank(val countOfMatch: Int, val winningMoney: Int) {
     THIRD(5, 1_500_000),
     FOURTH(4, 50_000),
     FIFTH(3, 5_000),
-    MISS(0, 0), ;
+    MISS(0, 0),
+    ;
 
     companion object {
         fun valueOf(
