+++ src/test/kotlin/lotto/model/LottoNumberTest.kt
@@ -1,4 +1,4 @@
-package lotto
+package lotto.model
 
 import org.junit.jupiter.api.assertThrows
 import org.junit.jupiter.params.ParameterizedTest
