+++ src/test/kotlin/lotto/model/LottoTest.kt
@@ -1,4 +1,4 @@
-package lotto
+package lotto.model
 
 import org.junit.jupiter.api.Test
 import org.junit.jupiter.api.assertDoesNotThrow
