+++ src/test/kotlin/lotto/model/PrizeAmountManagerTest.kt
@@ -1,4 +1,4 @@
-package lotto
+package lotto.model
 
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.api.Test
